+++ apps/imagine-api/src/photo/photo-reaction.resolver.ts
@@ -4,10 +4,14 @@ import {PhotoReactionModel} from './photo-reaction.model';
 import {HasSession} from '../session/has-session.decorator';
 import {GetSession} from '../session/get-session.decorator';
 import {PhotoReactionService} from './photo-reaction.service';
-import {PhotoReactionCreateInput} from './photo-reaction.input';
+import {Args, Mutation, Query, Resolver} from '@nestjs/graphql';
 import {PhotoFilterManyInput, PhotoFilterOneInput} from './photo.input';
 import {PhotoReactionDataloaderService} from './photo-reaction.dataloader';
-import {Args, Mutation, Query, Resolver} from '@nestjs/graphql';
+import {
+  PhotoReactionCreateInput,
+  PhotoReactionFilterOneInput,
+} from './photo-reaction.input';
+import {UnauthorizedException} from '@nestjs/common';
 
 @Resolver(() => PhotoReactionModel)
 export class PhotoReactionResolver {
@@ -36,11 +40,20 @@ export class PhotoReactionResolver {
 
   @Mutation(() => PhotoReactionModel)
   @HasSession()
-  async photoReactionCreate(
+  async photoReactionUpdate(
+    @Args('filter', {type: () => PhotoReactionFilterOneInput})
+    filter: PhotoReactionFilterOneInput,
     @Args('input', {type: () => PhotoReactionCreateInput})
     input: PhotoReactionCreateInput,
     @GetSession() session: UserEntity
   ): Promise<PhotoReactionModel> {
+    const matchingPhotoReaction = await this.photoReactionService.findOne({
+      id: filter.id,
+    });
+    const userOwnsPhotoReaction = matchingPhotoReaction.userID === session.id!;
+    if (!userOwnsPhotoReaction) {
+      throw new UnauthorizedException();
+    }
     return this.photoReactionService.create({
       userID: session.id!,
       reaction: input.reaction,
@@ -49,7 +62,15 @@ export class PhotoReactionResolver {
   }
 
   @Mutation(() => Boolean)
-  async photoReactionDelete(@Args('id') id: number) {
+  async photoReactionDelete(
+    @Args('id') id: number,
+    @GetSession() session: UserEntity
+  ) {
+    const matchingPhotoReaction = await this.photoReactionService.findOne({id});
+    const userOwnsPhotoReaction = matchingPhotoReaction.userID === session.id!;
+    if (!userOwnsPhotoReaction) {
+      throw new UnauthorizedException();
+    }
     await this.photoReactionService.delete(id);
     await this.photoReactionDataloaderService.clearByID(id);
     return true;
